fix union sub-types are not added to the union (#28)
This PR is about to fix an issue with union types, I stumbled over. Since v0.0.17 this was broken for me, because the union type appeared to be empty, at least for the introspection. 

I boiled down the issue to be introduced in PR #19. I guess it was a simple typo back then, since the if clause is currently decoupled from the loop.

I modified the test as well, to ensure the `union.Types` slice is actually populated